Introduced storage-layer abstraction.
This commit creates the new file `storage-api.go` which contains
an abstraction layer for the storage/retrieval of the data.

The intention is to allow others to present a different kind of
storage system - for example they might wish to store their data
in MySQL, Cassandra, Redis, or similar.

With this layer in-place that becomes both possible and simple.
There is no more mixing of the HTTP-layer stuff and the storage
details.

This closes #11.