Get/Set meta-data in addition to regular-data.
This updates #17 by changing the blob-servers' storage-API
to allow getting/setting meta-data (simple key=value) along
with the data itself.

You can test this by launching a blob-server and setting
some headers on the upload that have an X-prefix:

   deagol ~ $ curl -X POST -H "X-Test: me" -H "X-Data: $(date)" --data-binary @/proc/uptime http://localhost:3001/blob/meta
   {"id":"meta","status":"OK","size":22}

Now when you retrieve that blob the X-headers are returned:

    deagol ~ $ curl -v http://localhost:3001/blob/meta
    * Hostname was NOT found in DNS cache
    *   Trying ::1...
    * connect to ::1 port 3001 failed: Connection refused
    *   Trying 127.0.0.1...
    * Connected to localhost (127.0.0.1) port 3001 (#0)
    > GET /blob/meta HTTP/1.1
    > User-Agent: curl/7.38.0
    > Host: localhost:3001
    > Accept: */*
    >
    < HTTP/1.1 200 OK
    < X-Data: Fri May 27 07:39:10 EEST 2016
    < X-Test: me
    < Date: Fri, 27 May 2016 04:39:13 GMT
    < Content-Length: 22
    < Content-Type: text/plain; charset=utf-8
    <
    1622461.79 6266107.13

NOTE: As a special-case `X-Mime-Type` on upload will be returned as
`Content-Type` on download.  This is good.