Previously the upload-handler test only succeeded by accident.
We were testing /blob/foo-bar - which should have failed because "-"
doesn't match /[a-z0-9]/.  However the value wasn't being read from
the URL, so instead we received a failure because "" doesn't match
the regular expression.

Reworked the test of the upload-handler such that it actually tests
what it is supposed to do - and added a new test which uploads a
valid file to confirm it works.